requirements.txt
flake8-black==0.2.3
flake8-builtins==1.5.3
flake8-isort==4.0.0
ipdb==0.13.9
matplotlib==3.4.3
pandas==1.3.3
pep8-naming==0.12.1
pipdeptree==2.1.0
pytest-cov==2.12.1
scipy==1.7.1
sympy==1.8
wheel==0.37.0
